Financial Budgeting for TAMUhack 2022
This program was created for the hackathon hosted at Texas A&M University. 

Due to the lack of financial literacy classes available, this program helps you determine how much money you might be left with after essential expenses such as income tax, health/car insurance, groceries, rent etc., based on your inputted income and state of residency. This can come to use while switching jobs or moving to a new state. Eventually, this can be developed on a city basis as cities and towns within the same state have different costs of living.